update: core functions in different file
